requirements.txt
fastapi==0.104.1
motor==3.4.0
pymongo==4.6.1  # or the version that matches
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.0.1  # For image handling in tests
python-dateutil==2.8.2  # For date handling
cloudinary==1.36.0  # For cloud image storage